feat(source): Implemented json writer and reader for board
What changed:
- Modified pom file
- Service class for abstraction when retrieving json objects
- Implemented json reader and writer into the main method
