--- src/main/java/com/ntnu/idatt/MainApplication.java
@@ -1,16 +1,0 @@
-package com.ntnu.idatt;
-
-import javafx.application.Application;
-import javafx.stage.Stage;
-
-public class MainApplication extends Application {
-
-  public static void main(String[] args) {
-    launch();
-  }
-
-  @Override
-  public void start(Stage stage) {
-
-  }
-}
