--- src/main/java/com/ntnu/idatt/logic/action/SomeOtherAction.java
@@ -1,13 +1,0 @@
-package com.ntnu.idatt.logic.action;
-
-import com.ntnu.idatt.entity.Player;
-
-/**
- * placeholder for other actions that a player must do if landed on the tile containing the action
- */
-public class SomeOtherAction implements TileAction {
-
-  public void perform(Player player) {
-
-  }
-}
