--- src/main/java/com/ntnu/idatt/utils/InputValidation.java
@@ -1,5 +1,0 @@
-package com.ntnu.idatt.utils;
-
-public class InputValidation {
-
-}
