+++ src/main/java/edu/ntnu/bidata/idatt/MainApplication.java
@@ -1,0 +1,16 @@
+package edu.ntnu.iir.bidata;
+
+import javafx.application.Application;
+import javafx.stage.Stage;
+
+public class MainApplication extends Application {
+
+  public static void main(String[] args) {
+    launch();
+  }
+
+  @Override
+  public void start(Stage stage) {
+
+  }
+}
