+++ src/main/java/edu/ntnu/bidata/idatt/entity/Die.java
@@ -1,0 +1,18 @@
+package edu.ntnu.iir.bidata.entity;
+
+import java.util.Random;
+
+public class Die {
+
+  private final Random random = new Random();
+  private int lastRolledValue;
+
+  public int roll() {
+    lastRolledValue = random.nextInt(6) + 1;
+    return lastRolledValue;
+  }
+
+  public int getValue() {
+    return lastRolledValue;
+  }
+}
