+++ src/main/java/edu/ntnu/bidata/idatt/logic/action/SomeOtherAction.java
@@ -1,0 +1,13 @@
+package edu.ntnu.iir.bidata.logic.action;
+
+import edu.ntnu.iir.bidata.entity.Player;
+
+/**
+ * placeholder for other actions that a player must do if landed on the tile containing the action
+ */
+public class SomeOtherAction implements TileAction {
+
+  public void perform(Player player) {
+
+  }
+}
