+++ src/main/java/edu/ntnu/bidata/idatt/logic/action/TileAction.java
@@ -1,0 +1,8 @@
+package edu.ntnu.iir.bidata.logic.action;
+
+import edu.ntnu.iir.bidata.entity.Player;
+
+public interface TileAction {
+
+  void perform(Player player);
+}
