+++ src/main/java/edu/ntnu/bidata/idatt/service/BoardService.java
@@ -1,0 +1,4 @@
+package edu.ntnu.bidata.idatt.service;
+
+public class BoardService {
+}
