+++ src/main/java/edu/ntnu/bidata/idatt/utils/InputValidation.java
@@ -1,0 +1,5 @@
+package edu.ntnu.iir.bidata.utils;
+
+public class InputValidation {
+
+}
