+++ src/main/resources/edu/ntnu/bidata/idatt/players.csv
