--- src/test/java/com/ntnu/idatt/model/BoardTest.java
@@ -1,11 +1,0 @@
-package com.ntnu.idatt.model;
-
-import org.junit.jupiter.api.Test;
-
-class BoardTest {
-
-  @Test
-  void test() {
-
-  }
-}
