+++ src/test/java/edu/ntnu/bidata/idatt/model/BoardTest.java
@@ -1,0 +1,11 @@
+package edu.ntnu.iir.bidata.model;
+
+import org.junit.jupiter.api.Test;
+
+class BoardTest {
+
+  @Test
+  void test() {
+
+  }
+}
